requirements.txt
beautifulsoup4==4.11.1
bs4==0.0.1
mysqlclient==2.1.0
pandas==1.3.5
pytesseract==0.3.9
requests==2.26.0
